engine selection problem
git-svn-id: svn://svn.forge.objectweb.org/svnroot/spagobi/V_4.x/Server/trunk@22179 99afaf0d-6903-0410-885a-c66a8bbb5f81

+++ SpagoBIProject/src/it/eng/spagobi/tools/multitenant/service/rest/MultitenantCRUD.java
@@ -154,7 +154,7 @@ public class MultitenantCRUD {
 		
 		for (SbiOrganizationEngine orgEngine: selectedEngines) {		
 			SbiEngines sbiEngine = orgEngine.getSbiEngines();
-			if(sbiEngine.getEngineId() == engine.getId())
+			if(sbiEngine.getEngineId().intValue() == engine.getId().intValue())
 				return true;		
 		}
